+++ Demo/KeyboardCustom/DemoKeyboardView.swift
@@ -1,8 +1,8 @@
 //
 //  DemoKeyboardView.swift
-//  KeyboardCustom
+//  KeyboardKitDemo
 //
-//  Created by Daniel Saidi on 2022-09-02.
+//  Created by Daniel Saidi on 2022-02-04.
 //  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
 //
 
@@ -10,40 +10,23 @@ import KeyboardKit
 import SwiftUI
 
 /**
- This view sets up a `SystemKeyboard` as a keyboard view and
- adds an `AutocompleteToolbar` above it.
+ This view wraps a `SystemKeyboard`, which aims to mimic the
+ native iOS keyboard as closely as possible.
 
- This view automatically hides the autocomplete toolbar when
- the keyboard doesn't prefer autocomplete, but the height is
- still allocated, since it's needed by the keyboard callouts.
+ While the `SystemKeyboard` is very flexible, it can also be
+ easily created by just providing it with a controller, like
+ we do here. This configuration adds an autocomplete toolbar
+ above the keyboard and replaces the system keyboard with an
+ emoji keyboard whenever needed.
  */
 struct DemoKeyboardView: View {
 
-    var controller: KeyboardInputViewController
-    
-    @EnvironmentObject
-    private var autocompleteContext: AutocompleteContext
+    unowned var controller: KeyboardInputViewController
 
     @EnvironmentObject
     private var keyboardContext: KeyboardContext
 
     var body: some View {
-        VStack(spacing: 0) {
-            if keyboardContext.prefersAutocomplete {
-                autocompleteToolbar
-            }
-            SystemKeyboard(controller: controller)
-        }
-    }
-}
-
-private extension DemoKeyboardView {
-
-    var autocompleteToolbar: some View {
-        AutocompleteToolbar(
-            suggestions: autocompleteContext.suggestions,
-            locale: keyboardContext.locale,
-            suggestionAction: controller.insertAutocompleteSuggestion
-        ).opacity(keyboardContext.prefersAutocomplete ? 1 : 0)  // Still allocate height
+        SystemKeyboard(controller: controller)
     }
 }